+++ src/components/form/index.d.ts
@@ -1,5 +1,6 @@
 import { CommonProps } from '../common';
 /// <reference path="./checkbox/index.d.ts" />
+/// <reference path="./field_empty/index.d.ts" />
 /// <reference path="./field_number/index.d.ts" />
 /// <reference path="./field_password/index.d.ts" />
 /// <reference path="./field_search/index.d.ts" />