+++ src/components/form/index.js
@@ -3,6 +3,7 @@ export {
   EuiCheckboxGroup,
 } from './checkbox';
 export { EuiDescribedFormGroup } from './described_form_group';
+export { EuiFieldEmpty } from './field_empty';
 export { EuiFieldNumber } from './field_number';
 export { EuiFieldPassword } from './field_password';
 export { EuiFieldSearch } from './field_search';