+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `text` as a color option for `EuiLink` ([#1571](https://github.com/elastic/eui/pull/1571))
 - Added `EuiResizeObserver` to expose ResizeObserver API to React components; falls back to MutationObserver API in unsupported browsers ([#1559](https://github.com/elastic/eui/pull/1559))
+- Added `EuiFieldEmpty` to align data text with inputs in inline forms
 
 **Bug fixes**